+++ e2e/tests/plexSearch.spec.ts
@@ -127,7 +127,7 @@ test("Plex Search: Should hide 'Search on Plex' button when PLEX var does not ex
         auth: {
           token: "mock-token",
           refresh_token: "mock-refresh-token",
-          expires: 1234567890,
+          expires_at: 1234567890,
           user_id: "192283714",
           country_code: "FR",
         },